Simplify price integration tests
