+++ tests/NHSD.GPIT.BuyingCatalogue.E2E.PublicBrowseTests/Areas/Admin/ListPrices/AdditionalService/Tiered/EditTierPrice.cs
@@ -1,21 +1,16 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-using FluentAssertions;
-using NHSD.GPIT.BuyingCatalogue.E2ETests.Framework.Objects.Admin.ListPrices;
+﻿using System;
+using System.Collections.Generic;
+using NHSD.GPIT.BuyingCatalogue.E2ETests.Areas.Admin.ListPrices.Base.Tiered;
 using NHSD.GPIT.BuyingCatalogue.E2ETests.Utils;
-using NHSD.GPIT.BuyingCatalogue.E2ETests.Utils.TestBases;
 using NHSD.GPIT.BuyingCatalogue.EntityFramework.Ordering.Models;
-using NHSD.GPIT.BuyingCatalogue.Framework.Extensions;
 using NHSD.GPIT.BuyingCatalogue.WebApp.Areas.Admin.Controllers;
-using Xunit;
 
 namespace NHSD.GPIT.BuyingCatalogue.E2ETests.Areas.Admin.ListPrices.AdditionalService.Tiered
 {
-    public sealed class EditTierPrice : AuthorityTestBase, IClassFixture<LocalWebApplicationFactory>
+    public sealed class EditTierPrice : EditTierPriceBase
     {
         private const int CataloguePriceId = 14;
         private const int TierId = 17;
-        private const int TierIndex = 1;
         private static readonly CatalogueItemId SolutionId = new(99998, "001");
         private static readonly CatalogueItemId AdditionalServiceId = new(99998, "001A99");
 
@@ -27,54 +22,16 @@ namespace NHSD.GPIT.BuyingCatalogue.E2ETests.Areas.Admin.ListPrices.AdditionalSe
             { nameof(TierId), TierId.ToString() },
         };
 
-        private static readonly Dictionary<string, string> QueryParameters = new()
-        {
-            { nameof(TierIndex), TierIndex.ToString() },
-        };
-
         public EditTierPrice(LocalWebApplicationFactory factory)
             : base(
                 factory,
                 typeof(AdditionalServiceListPriceController),
-                nameof(AdditionalServiceListPriceController.EditTierPrice),
-                Parameters,
-                QueryParameters)
+                Parameters)
         {
         }
 
-        [Fact]
-        public void AllSectionsDisplayed()
-        {
-            using var context = GetEndToEndDbContext();
-            var catalogueItem = context.CatalogueItems.Single(p => p.Id == AdditionalServiceId);
+        protected override Type Controller => typeof(AdditionalServiceListPriceController);
 
-            CommonActions.PageTitle().Should().Be($"Edit Tier {TierIndex} price - {catalogueItem.Name}".FormatForComparison());
-            CommonActions.LedeText().Should().Be("Provide information about this tier.".FormatForComparison());
-
-            CommonActions.GoBackLinkDisplayed().Should().BeTrue();
-            CommonActions.SaveButtonDisplayed().Should().BeTrue();
-
-            CommonActions.ElementIsDisplayed(EditTierPriceObjects.PriceInput).Should().BeTrue();
-        }
-
-        [Fact]
-        public void ClickGoBackLink_NavigatesToCorrectPage()
-        {
-            CommonActions.ClickGoBackLink();
-
-            CommonActions.PageLoadedCorrectGetIndex(
-                typeof(AdditionalServiceListPriceController),
-                nameof(AdditionalServiceListPriceController.EditTieredListPrice)).Should().BeTrue();
-        }
-
-        [Fact]
-        public void ClickSave_NavigatesToCorrectPage()
-        {
-            CommonActions.ClickSave();
-
-            CommonActions.PageLoadedCorrectGetIndex(
-                typeof(AdditionalServiceListPriceController),
-                nameof(AdditionalServiceListPriceController.EditTieredListPrice)).Should().BeTrue();
-        }
+        protected override CatalogueItemId CatalogueItemId => AdditionalServiceId;
     }
 }
